updated after long time

+++ GUI/MapViewer-API/tester_mapviewer.cpp
@@ -5,7 +5,7 @@ Tester_MapViewer::Tester_MapViewer(QObject *parent) : QObject(parent)
 
 }
 
-void Tester_MapViewer::send_Data(double longitude, double latitude)
+void Tester_MapViewer::send_Data(float longitude, float latitude)
 {
     emit incoming_Data(longitude,latitude);
 }
